+++ sample-apps/discount-functions-sample-app/web/frontend/hooks/useRedirectToDiscounts.ts
@@ -2,13 +2,13 @@ import { useAppBridge } from '@shopify/app-bridge-react';
 import { Redirect } from '@shopify/app-bridge/actions';
 import { useCallback } from 'react';
 
+const APP_DISCOUNTS_PATH = '/discounts/apps';
+
 export function useRedirectToDiscounts() {
   const app = useAppBridge();
 
   return useCallback(() => {
     const redirect = Redirect.create(app);
-    redirect.dispatch(Redirect.Action.ADMIN_SECTION, {
-      name: Redirect.ResourceType.Discount,
-    });
+    redirect.dispatch(Redirect.Action.ADMIN_PATH, APP_DISCOUNTS_PATH);
   }, [app]);
 }
