--- sample-apps/discount-functions-sample-app/web/frontend/hooks/useShopifyMutation.js
@@ -1,17 +1,0 @@
-import { useMutation } from "react-query";
-import { GraphQLClient } from "graphql-request";
-
-import { useAuthenticatedFetch } from "./useAuthenticatedFetch.js";
-
-export const useShopifyMutation = ({ query }) => {
-  const authenticatedFetch = useAuthenticatedFetch();
-  const graphQLClient = new GraphQLClient("/api/graphql", {
-    fetch: authenticatedFetch,
-  });
-
-  const { mutateAsync, ...mutationProps } = useMutation(async (variables) =>
-    graphQLClient.rawRequest(query, variables)
-  );
-
-  return [mutateAsync, mutationProps];
-};
